+++ FarmerzonArticlesDataAccessModel/Unit.cs
@@ -5,7 +5,7 @@ namespace FarmerzonArticlesDataAccessModel
     public class Unit
     {
         // primary key
-        public int UnitId { get; set; }
+        public long UnitId { get; set; }
         
         // relationships
         public IList<Article> Articles { get; set; }
