+++ FarmerzonArticlesDataTransferModel/Unit.cs
@@ -4,7 +4,7 @@ namespace FarmerzonArticlesDataTransferModel
 {
     public class Unit
     {
-        public int UnitId { get; set; }
+        public long UnitId { get; set; }
         
         public IList<Article> Articles { get; set; }
